@@ -12,12 +12,19 @@ namespace MauiApp1.Views
             {
                 Children =
                 {
+                    // Typed binding
+                    
                     new Label() { Text = "Typed Binding:" },
-                    new Picker().Bindv2(Picker.ItemsSourceProperty, static (MainViewModel vm) => vm.Numbers)
-                                .Bindv2(static (MainViewModel vm) => vm.Index),
+                    // ItemSource uses the package provided method
+                    // SelectedIndex uses local definition
+                    new Picker().Bind(Picker.ItemsSourceProperty, static (MainViewModel vm) => vm.Numbers)
+                                .BindV2(static (MainViewModel vm) => vm.Index),
+
+                    // String-based
+
                     new Label() { Text = "Classic Binding:" },
                     new Picker().Bind(Picker.ItemsSourceProperty, nameof(MainViewModel.Words))
-                                .Bind(nameof(MainViewModel.Index), BindingMode.TwoWay)
+                                .Bind(nameof(MainViewModel.Index)) // It's two-way by definition
                 }
             }.Center();
         }
